server/requirements.txt
fastapi[standard] == 0.116.*
pydantic-settings == 2.10.*
python-telegram-bot